Fixes to AudioBuffer, Beginning Red::Graphics::Laminar (2D Graphics layer), Red::Audio::Effects::Echo
Fixed a problem with AudioBuffer that would cause the effects of
ClearBufferFloat to be truncated too early. Began Laminar, a 2d sprite
graphics layer on top of Xenon, and finished implementing
Red::Effects::Echo.

+++ include/Red/Audio/IAudioOutputDevice.h
@@ -30,7 +30,7 @@ namespace Red
 			virtual void Stop () = 0;
 			
 			virtual uint32_t GetSampleFrequency () = 0;
-			virtual uint32_t GetSampleSize () = 0;
+			virtual uint64_t GetSampleSize () = 0;
 			virtual uint32_t GetChannelCount () = 0;
 			
 			virtual const Util::Time :: Duration GetFillDuration () = 0;
